Move transcription check sending to slack helpers
